Move benchmark tests into upper folder
