+++ test/e2e/metric_util.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package kube_batch
+package e2e
 
 import (
 	"bytes"
